[RENAME]  : ✏️ cookie 파일 위치 변동에 따른 import 위치 변동
